Incorporating feedback from pull request.
- Created abstract parent class AbstractStreamDataSource.
  Its purpose is to handle all the functionality for streaming
  input sources across file input sources and non-file input sources.
