+++ hydra-task/src/main/resources/hydra-sources.conf
@@ -15,6 +15,7 @@ plugins {
       markDir: marks
     }
     prefetch: DataSourcePrefetch
+    query: DataSourceQuery
     range: DataSourceRange
     sorted: SortedTaskDataSource
   }
@@ -49,16 +50,18 @@ com.addthis.hydra.task {
   }
 
   source {
+    AbstractStreamDataSource {
+      buffer: 128
+      pollInterval: 1 second
+      pollCountdown: 1800
+    }
     AbstractStreamFileDataSource {
       magicMarksNumber: 42
       workers: 2
-      buffer: 128
       skipSourceExit: 0
       multiBundleReads: 1
       preOpen: 1
       format: {channel {}}
-      pollInterval: 1 second
-      pollCountdown: 1800
       latchTimeout: 30 seconds
 
       # support legacy system property paths for now
@@ -77,9 +80,6 @@ com.addthis.hydra.task {
       params: {}
     }
     DataSourceQuery {
-      buffer: 128
-      pollInterval: 1000
-      pollCountdown: 1800
       treeDir: "data"
     }
   }
